Adds notion of ColorProviders class for colors in Glance.
Adds a Material 3 inspired ColorProviders class.
This CL makes use of it with an implementation of dynamic coloring,
initially used only internally for templates, but will be expanded once
themeing is more fleshed out.
This is part of a bigger effort for full themeing.

The color values need to be in their own files, rather than using the
material 3 theme directly, as the color res ids are resolved in the app
process, not the launcher process.
The binding for day / night needs to happen when launcher switches, not
when the Glance is made.

Bug: 237012816
Test: TemplateTests
RelNote: Adds a ColorProviders class that can be used as part of themes for Glance.

Change-Id: I848b91655b34e5542551e9c581bc72097229f92c