Add pointer checks
Bug: 233209880

There are few possibilities for how Offset.getX can cause
a failure at the line in the bug. It could be caused by
an infinite value in both X & Y coordinates and the Matrix
has a 0 value or the opposite. It could also be that the
matrix calculation ends up with a NaN value in one of
the cells.

This CL adds checks to ensure that the Matrix's map()
 operation doesn't generate a NaN. It also
uses the platform's transformMatrixToGlobal() instead of
the local implementation when it is available. Finally,
it ensures that if NaN or Infinite values are passed
in the MotionEvent or the View isn't attached,
then it isn't processed.

I cannot replicate the bug, so I'm trying to find all
possible sources for this error. It may be that the
error only exists on some hardware.

Test: new tests.
Change-Id: I56c9483305be77c49cdc5d10601955b84af0f288