Add Kruth declarations required for commonizing datastore-core tests
Adding minimal set of Truth conversions (Kruth) required to support
datastore-core tests. The goal here is to emulate Truth's API surface in
a KMP-compatible way, that allows us commonize projects like datastore
without changing the source of the tests.

This also removes the Truth dependency as a first step in commonizing
datastore-core tests. We still need a good solution for Rules and
TemporaryFiles, but will follow up on those in a separate change.

Note: The Kruth assertions are never meant to be published and are
deprecated from inception. Instead of using such a shim, we should
eventually just migrate to use a common assertion framework directly,
but this allows us to convert libraries to KMP with more confidence.

Test: ./gradlew :datastore:datastore-core:test
Test: ./gradlew :internal-testutils-kmp:allTests
Change-Id: I342ead724f75fcdb7652b6b411f52436fa7cdc2a