Fix AmbientMode tests for Wear.
Previously, these tests were instrumentation tests, running on a phone
emulator. Because of that, the phone emulator didn't have the Wearable
library (of course), so WearableActivityController would be used from
the test instead. Now that we have a Wear emulator though, the Wear
runner _does_ have the Wearable library, and because the main library
has a <uses-library> stanza in its Manifest for this library,
WearableActivityController will instead be loaded from the platform.
This platform version does not have the static getLastInstance call
(since it's a test shim), leading to test failures on that runner.

In reality, those tests do not need to be instrumentation tests; they
can run perfectly fine as unit tests. Migrated to be unit tests, and
fixed.

This also migrates all the ActivityScenarioRules to be a resolved
ActivityScenario. Robolectric requires all launced Activities to be
declared in a manifest, but AGP does not merge manifests from tests.
Instead, the launchActivity helper registers the Activity, then launches
it using ActivityScenario. We can't use this as a rule any more (with
the activity registration in @Before), since rules seem to run before
@Before, so it attempts to launch an unregistered activity.

Bug: 219049525
Test: ./gradlew :wear:wear:test
Change-Id: I61ed01eecbc9dc9c2b2a384dcd1f19ed79998c96