Expose RoomDatabase getters via properties.
When RoomDatabase was in Java, Kotlin callers could use the property syntax to get some of the fields, `openHelper`, `queryExecutor`, etc. By moving Room to Kotlin and keeping the getters as function and the fields internal or private, we break source incompatibility making the migration harder. This changes makes it so that we use properties, keeping ABI compatibility and source compatibility at the cost of extra internal properties.

Relnote: "Fix a Kotlin source incompatible change for getter / properties of getOpenHelper, getQueryExecutor and getTransactionExecutor."
Bug: 234152465
Test: Existing
Change-Id: Iad0ac5775730c80127f57fe6eec58796c0aff5d6