Update the easing params on ScalingParams in line with UX Spec changes
There is a small change in the UX spec for easing on the scaling at the edges of a ScalingLazyColumn.kt - update the defaults to match.

Bug: 234123876
Test: ./gradlew :wear:compose:compose-material:connectedCheck --info --daemon
RelNote: "We have made a small adjustment in the ScalingLazyColumn easing as items reach the edge of the screen to match UX spec updates. Old values CubicBezierEasing(0.25f, 0.00f, 0.75f, 1.00f) -> new values CubicBezierEasing(0.3f, 0f, 0.7f, 1f). To keep the old behaviour you can override the scalingParams of ScalingLazyColumn"
Change-Id: Ie375ca83655ae09293b6565174787eac1e6b2e3e