Remove fillMaxWidth() from ListHeader.kt
ListHeader is setting fillMaxWidth() but it should behave the same way as Chip() and wrap its contents.

Note also tidied up a few ScalingLazyColumn samples and demos by adding an explicit .fillMaxWidth(). These samples where missed when removing previous fillMaxWidth() on ScalingLazyColumn and Chip and were resulting in demos and samples that didn't behave well as their contents changed and would fit the new contents and then fit to the left hand side of the screen.

Bug: 235074035
Test: ./gradlew :wear:compose:compose-material:connectedCheck --info --daemon
RelNote: "We have removed the explicit call to fillMaxWidth() on list header as it is not needed and can result in problems if a ScalinglazyColumn has a mixture of ListHeader() and Chip() components as the width will tend to grow/shrink as ListHeader items are scrolled in/out of view."
Change-Id: I3714408eac2c75cd8168162ec94a600b0b87f024