Support ambient before SysUI has started
Post R we can rely on Intent.ACTION_SCREEN_OFF and interrogate the
ambient_enabled global setting to determine if the system has
gone ambient. We do this in tandem with the signals sent from
SysUI (which tends to send the same thing albeit slightly delayed).

If SysUI has not called setWatchUiState, in ambient mode we now
assume SysUI isn't able to call ambientTickUpdate and we rely on
Intent.ACTION_TIME_TICK to drive ambient ticks. This fixes
ambient in the DirectBoot scenario.

In addition this patch simplifies BroadcastsReceiver multiplexing
all the intents onto a single receiver.

Bug: 234116384
Test: Added new unit tests. Also commented out InteractiveWatchFaceImpl.setWatchUiState & ambientTickUpdate and observed ambient transition and updates once per minute as expected.

Change-Id: Ib014528f39c7269217280fc505c181924f9c924e