Fixed ANR issues with GLRenderer
Fixed issue where blocking callbacks to destroy
surfaces in Surface/TextureView would not be fired
if there were multiple RenderTarget#detach calls
made concurrently. In this case the RenderTarget
would be fully detached before subsequent callbacks
would be invoked

Fixed issue where GLRenderer#stop callbacks would not
be invoked if a call to RenderTarget#requestRender
was never made. This was because EglManager is
lazily initialized on the first call to requestRender
and GLRenderer#stop would only release resources if
the EglManager instance was non-null.

Fixes: 235174482
Test: Added GLRendererTests
Change-Id: I7fd260ec541dc77f1c40c273ece5055ceb6e9ea4