Create fixes for class verification failures with unusual receivers
Instead of checking the type of the call receiver, this uses `UElement.asSourceString()` to generate the receiver for the wrapper call. A comment in the code said to consider this approach if no fix was generated often. Based on b/206131626, I tried using `this` and `super` as receivers. Handling those cases as well as casts would require cases for:
- JavaUThisExpression
- JavaUSuperExpression
- JavaUTypeCastExpression (which itself contains a UExpression that needs to turn into a string)

I also did a bit of experimenting with somewhat wacky receivers and some more cases that are possible are:
- JavaUCallExpression
- JavaUCompositeQualifiedExpression
- JavaUTernaryIfExpression
- JavaConstructorUCallExpression
And I’m sure there are more (I didn’t include tests for the cases above, but I could add them).

Based on this I think using `UElement.asSourceString()` makes sense, which is also how the arguments to the call are being generated.

Test: Added tests for call to cast object, as well as calls on this and super.
Bug: 206111383
Change-Id: Ic78dce4619ab70ffa59088b8e85861cffc8478da