Add Navigation Component
Add navigation dependency. It is the Compose navigation dependency which is required for
components such as the NavHostController in Compose.

Add Compose Icon dependency. This dependency allows us to access icons related to the camera.

Add ComposeCameraScreen enum. This enum denotes each destination we will have in the navigation graph.
We will use `Screen.name` as the route for navigation and each enum
will hold an Icon to display in the navigation tabs.

Add Navigation package which contains ComposeCameraScreen.kt and ComposeCameraNavHost.kt which contains a custom NavHost

Bug: 234080273
Test: n/a
Change-Id: Ie7150294ab4bc208e949152161f10736591c9aee