Add PermissionsUI and setup Screens
PermissionsUI is a wrapper around the content we wish to render. Each
time the app renders, PermissionUI runs the check to see if permissions
are granted. If yes, display content. If not, display a response to the
user.

Replace individual screens in NavHost. The NavHost destinations are replaced with the individual screens for
ImageCapture, VideoCapture and Gallery.

Bug: 234080273
Test: n/a
Change-Id: I1838728387ec96700a8a87a62d1baefcf33499e2