Disable KMP plugin klib cleanup service
This CL disables the klib cleanup service in the KMP gradle
plugin which eagerly deletes local klibs before dependencies
are fully resolved.

That service is already removed on the main branch of the plugin
and this CL simply disables the service so that we can run
`./studiow k` without disabling configuration cache (which was
another workaround).

Bug: 236850628
Test: update studio to build 8762557, observe ArraySetTest has no
red lines.

Change-Id: If57d74683e1650c82174ea226168f6fc42507e38