Fix failed tests on Windows
This fixes failed tests on Windows due to different file path format.
On Linux/Mac the path to a source file could be `/foo/bar/Baz.kt`,
while on Windows it could be `C:\foo\baz\Baz.kt`.

To verify the fix we extracted room-compiler-processing and
room-compiler-processing-testing into a standalone project,
removed AndroidX plugin, and made it build on Windows.

Test: Passed existing tests with the manually created project on Windows
Bug: b/230830024
Change-Id: Ia2640f30a67bff2df3388d23b0684358de125024