Fix issue with XAnnotationValues of lists
This CL fixes an issue with the XAnnotationValue#has*Value when dealing
with individual annotation values in an array annotation value. In
particular, we were incorrectly using the original method's return value
to deternmine the type of the individual annotation values, which
incorrectly resulted in the individual annotation values also being
categorized as list types (e.g. `int[]` instead of `int`).

Test: XAnnotationValueTest
Change-Id: Ibfb8c7d0ae791a6c4af8117de65729e97013848c