Make NavBackStackEntry arguments immutable
NavBackStackEntry arguments should always equal the arguments
created by the navigate() call. Unfortunately, Bundle is
inherently mutable, so to actually enforce immutability,
NavBackStackEntry now only sends a copy of the immutable
Bundle of arguments when you access the arguments property.

Relnote: "NavBackStackEntry's `arguments` and the
`arguments` passed to an `OnDestinationChangedListener`
are now only a copy of the immutable arguments that were
created when you navigated to the destination. This means
any changes made to those Bundles will not be reflected
in subsequent accesses to the `arguments` or other
`OnDestinationChangedListener` instances."
Test: new NavControllerTest passes
BUG: 220892847

Change-Id: I676f53df9aa311d51c5da1ae11d57b51bb149b7d