Add Integration Test for ComposeCameraApp
Added integration tests to ensure that PreviewView is streaming after
the Composables render. Currently there are two tests.

The first test ensures that PreviewView is streaming after the
Composable (ImageCaptureScreen) is rendered when the Activity launches.

The second test ensures that PreviewView is streaming when we navigate
from ImageCaptureScreen to VideoCaptureScreen using Compose navigation
framework.

Bug: 241035735
Test: ./gradlew camera:integration-tests:camera-testapp-uiwidgets:connectedAndroidTest
Change-Id: Iecee3d9e2f4171489534d4cb22e2b62af2b334ee