Introduce SyncFenceCompat onFirstShow strategy within HardwareBufferRenderer
This CL introduces onFirstShow, which will only send in a
SyncFenceCompat when the front buffer is transitioning from an invisible
to visible state. On subsequent renders, this will also create a sync
object if the front buffer usage bit is NOT supported by the device.
Otherwise we will create, peak, and discard the object.

Relnote: "SyncStrategy onFirstShow implemented for front buffered
rendering SyncFence utilization."

Bug: b/239076524
Test: ./gradlew :graphics:graphics-core:connectedAndroidTest
Change-Id: I1a67eef05e76abecdd9b860e60687fa23cb07d0b