Preview effects API review
This CL adds a new API for applying Preview effects. This is an advanced API for systematic users who are experts in Android graphics. For details, see design doc: go/camerax-effects-api

This CL exposes the API for Preview only. VideoCapture and other effects will be added once they are ready. The plan is for the API stays experimental for 1.2, and will become official in 1.3

Relnote: add a new API for applying post-processing effect to Preview output.

Bug: 240608986
Test: manual test and ./gradlew bOS
Change-Id: Ic17d53c89e7f03d9fc051334cba789209eb4d4ee