Use fully qualified class names in class verification failure autofix
This is a similar issue to aosp/2120172, but that only fixed the call receiver type.

Unit tests and studio shorten names differently, and for the two tests updated in this CL the previous behavior is better and still what happens in studio.
For one test (with java.lang.Class), the new test behavior is also correct. For the other (with drawable.Drawable), the new test behavior is incorrect: b/241573146. I'll file a separate bug to look at the test again for after the lint bug is responded to.

Test: Replaced existing unit test for qualified receiver type to also use qualified type for param and return.
Bug: 236721202
Change-Id: I73c2d7cf78e8e568b6cee11258ac6a09061ab300