Replace LIBRARY_GROUP_PREFIX in wear watchface-*
A number of APIs where marked with @RestrictTo(androidx.annotation.
RestrictTo.Scope.LIBRARY_GROUP_PREFIX) this is a mistake, they
should have been marked with @RestrictTo(androidx.annotation.
RestrictTo.Scope.LIBRARY_GROUP) instead. Technically this is a
breaking API change, however we're not aware of any uses of these
APIs outside of the library group in androidx so we think this is
OK.

Test: N/A
Bug: 241928149
Relnote: A number of APIs in wear watchface-* were incorrectly marked as RestrictTo.Scope.LIBRARY_GROUP_PREFIX, these are now correctly marked with RestrictTo.Scope.LIBRARY_GROUP. This has no effect on the public API surface.
Change-Id: I4dfc20846e42939ce9afd5429e7397cb2b2a80fc