Add WorkerParameters.getGeneration
Relnote: """
`WorkerParameters.getGeneration()` was added that returns
a generation of a worker. A worker has multiple generations, if it was updated via
`WorkManager.updateWork` or `WorkManager.enqueueUniquePeriodicWork` using
`ExistingPeriodicWorkPolicy.UPDATE`.
This worker can possibly be of an older generation rather than latest known,
if an update has happened while this worker is running."""
bug: 219446409

Test: WorkUpdateTest#updateWorkerParameterGeneration
Change-Id: I128a9838067f3a1722a9460f52ef58fb0f98136c