Add Renderer.sendPreviewImageNeedsUpdateRequest
Test: New test
Bug: 235607987
Relnote: We've added Renderer.sendPreviewImageNeedsUpdateRequest which is useful for watch faces that have state outside of the UserStyleSchema which affects the way they look (e.g. a watch face with a selectable background image). On the client side we've added PreviewImageUpdateRequestedListener as an optional paramater to getOrCreateInteractiveWatchFaceClient to observe these requests.
Change-Id: Iff44af72ef6d22973fd63137dcc6e4d02f6bff12