Fix PreviewViewDeviceTest#previewViewClicked_tapToFocusInvokedOnController failures
Some devices might have incorrect MotionEvent#getEventTime() and MotionEvent#getDownTime() info when issuing click event via UiDevice in the first time. The system time only occupies around 100ms but the event time difference is around 1000ms and then the click event is incorrectly recognized as long-click. The issue causes PreviewView to ignore the clickoperation and then onTapToFocus event can't be received.

Use a ClickEventHelper to monitor the incorrect long click event and retry the click operation to run the test.

Bug: 241206756
Test: Run HM local script to verify on test lab devices
Change-Id: Ic16737f90bb7566243ff5f614945c8bcc8855940