Add key parameter to hiltViewModel()
Add key parameter to hiltViewModel() to support creating multiple
instances of the same view model type.

Since Metalava complains that adding a parameter with default value
to an inline function is not binary compatible, we suppress it for
now: b/245139957.

Test: HiltViewModelComposeTest.kt
Relnote: Add key parameter to hiltViewModel() to support creating multiple instances of the same view model type.

Change-Id: I6ee4727a4b690bc35af4280b61ba2986d9eebb9b