Fix double nested graph popUpTo behavior
Navigation should always maintain the path to current destination as
part of that back stack. Under normal circumstances following the
recommended behavior this is true.

It is possible to create a scenario where this is not the case.

Given the following graph:

           A
	   |
	   B
	  / \
	 C   D
	     |
	     E
	     |
	     F

Where C is the start destination of graph B, if you navigate to D while
popping C off the back stack (which is not recommended), Navigation
fails to build the proper hierarchy for D resulting in B being popped
incorrectly and causing an IllegalStateException.

We should ensure that Navigation properly handles nested destinations.

RelNote: "Fixed `IllegalStateException` caused by navigating to a double
nested graph that shares a parent with a new popped start destination."
Test: added testLifecycleDoubleNestedGraph
Bug: 243778589

Change-Id: I9f7cb2445e6021c6ab4cb81d62612411249c1257