Fix the z-order of the items in Compose for Wear OS Scaffold
The z-order of the items in the Compose for Wear OS scaffold is wrong. The indicators (position and page) should sit on top of the vignette.

Bug: 244207528
Test: ./gradlew :wear:compose:compose-material:connectedCheck --info --daemon
Relnote: "We have corrected the z-order of position and page indictors in the Scaffold. The indicators will now sit on top of the vignette and so will not be obscured by the vignette if it is present."
Change-Id: Ib988f8f73d99fc2cfe5d29d0ea0198e68c391a18