Fix SurfaceFlinger OOM issue.
Fixed issue where new SurfaceControl instances
were being created on each render to the front
buffer instead of reusing the same SurfaceControl
instance. Updated the buildReparentTransaction
implementation to always pass the parentSurfaceControl
instead of the SurfaceView directly. This was because
the implementation of SurfaceControlCompat.Transaction#
reparent that consumed a SurfaceView internally created
a SurfaceControl each time.

Relnote: "Removed SurfaceControlCompat.Transaction
reparent API that consumed a SurfaceView as it internally
created SurfaceControl instances that were never released
and consuming resources indefinitely."

Fixes: 244489658
Test: Added test to GLFrontBufferedRenderer test
Change-Id: Ib7ff25337ef763c219045780a48101c43770cf19