Add Modifier.scrollAway for scrolling TimeText out of view.
Modifier.scrollAway is inspired by the Modifier.fadeAway family of Modifiers from Horologist.

Bug: 240535543
Test: Added to androidx.wear.compose.integration.demos
Relnote: "We have added Modifier.scrollAway, which scrolls an item vertically in/out of view, based on scroll state (with overloads to work with Column, LazyColumn and ScalingLazyColumn. ScrollAway is typically used to scroll a TimeText out of view as the user starts to scroll a list of items upwards."
Change-Id: I61766c2fa771df19a0accbd61b04cd22f6ecd7a2