Dispatch config change on added fragmentManagers
With the new provider callbacks, each fragment manager registers for
config change with the host once it is attached, and does not unregister
until it is destroyed. This means that even after fragments have been
put on the back stack, their childFragmentManagers can still dispatch
config changes, even though the fragment is not currently added.

We need to make it so that only added fragments (and their child
fragment managers) dispatch config changes.

RelNote: "Fixed an error that caused fragments on the back stack  to get
onConfigurationChanged() callbacks."
Test: Added ProviderCallbackTest
Bug: 242570955
Fixes: 243982866

Change-Id: I8dfe68f7b8700f0c1d3b13a3d5b826fc924b7452