Only setup MeteringRepeating when unattached
*) The shouldAddRepeatingUseCase() returns
True if there's no MeteringRepeating in both attached
and active states.
*) The shouldAddRepeatingUseCase() will be called
twice when it is attached, the first time is when the
MeteringRepeating is attached.
The second time is in the
attach(useCases: List<UseCase>) method,
it calls refreshAttachedUseCases before the
MeteringRepeating is added into the activated list.

As a result, addRepeatingUseCase() will be called twice,
and the setup of the MeteringRepeating will be duplicated.
It causes the configured DeferrableSurface be closed
accidentally.

This CL will update the shouldAddRepeatingUseCase() to
return False when MeteringRepeating is attached.

Bug: 246753647
Test: ImageCaptureTest
Change-Id: I7159e963d3bcd43cab3545e1970abc0515c3a702