Update the root layout index when an element loses an action
This change updates the LayoutNode proto so that it records whether or
or not an element has an action not. This causes the root layout ID to
be incremented when an element loses/gains an action, which prevents
recycling the view hierarchy.

Updated GlanceAppWidget to use LayoutConfiguration for T+ (addView).

unsetAction() and doNotUnsetAction() are removed since they are no
longer necessary.

Bug: 242397933
Test: GlanceAppWidgetReceiverTest#unsetActionCallback and
 GlanceAppWidgetReceiverTest#unsetCompoundButtonActionCallback
Change-Id: I6db34d58c4fb65160273362dda742067be4f2f87