Convert more ColumnTypeAdapters to XPoet
* Boxed primitives
* String
* Boolean (which is really primitive boolean to int and its boxed version)
* ByteArray (blob)

Updated the signature of CustomTypeConverter but its not actually fully converted to XPoet.

Since KSP has nullability information and nullability is part of Kotlin's type system, core adapters such as String and ByteArray will not generated null-check code is type is known to be not-null. This diverges the generated code between javac and ksp and tests were updated to reflect that.

Bug: 127483380
Test: Existing tests + KotlinCodeGenTest
Change-Id: Ic3d06ac7a12fb6d2e55de60e1b75270a4541217d