Migrate the simple @Insert writer path to XPoet.
Specifically InstantInsertMethodBinder -> InsertOrUpsertMethodAdapter -> EntityInsertionAdapterWriter. By migrating a WRITE query we are now able to verify the various other adapters and their bind statement path.

A notable change is that the shared enum converter functions will no longer handle nullable enums / strings, instead the
cursor / field null check is done before calling the converter function. This makes the Kotlin codegen easier as it does not have to add `checkNotNull` functions and it aligns with the rest of the adapters.

An API change too: The SupportSQLiteStatement parameter in EntityInsertionAdapter.bind() was incorrectly nullable instead of non-null, this is now fixed. However this is technically an incompatible nullability change but it should be safe for us as the class is restricted and only implemented by Room's generated code.

New XPoet APIs:
* XTypeName.getArrayTypeName - For representing arrays, including primitive arrays.

Test: Existing + KotlinCodeGenTest
Change-Id: I2e5e871523ee131f107105eb2783877ba5d724b1