Hold the usage of the DeferrableSurface during configuration
The DeferrableSurface may be closed before receiving the
`active` signal from the CameraGraph. The reason is CameraX
doesn't set the DeferrableSurface is `in use` during
configuring the CameraGraph.

This CL will set the used count on DeferrableSurface
during configuring, to ensure the Surface is valid
until the configuration is completed.

Bug: 251296842
Test: ImageCaptureTest
Change-Id: I70ca19e1587d50c45cd46a36c39fcb696c7ab8c6