When aborting TakePictureRequest, also abort the requests being processed.
When TakePictureManager is aborted, we need to abort all the callbacks that are not recieved by the app. Currently, we are ignoring the requests in the post-processing stage. In this CL, we track all the requests that are not recevied by the app, and abort them when TakePictureManager#abortAllRequest is called.

- In RequestWithCallback, replace the mIsComplete flag with a mCompleteFuture. The Future completes when the request gets an result or an error.
- In TakePictureManager, rename mInFlightRequest to mCapturingRequest to be more descriptive. Add a new mIncompleteRequests to track all the unfinished business.

Bug: 249593716
Test: manual test and ./gradlew bOS
Change-Id: Id4dd16c48bc0d9535ba0977c464ad0b4d74594ab