[Duo/Meet][Foldable] Fix the IllegaArgumentStateException when folding the device and unfold quickly in a meeting state
DisplayInfoManager.getMaxSizeDisplay only returns the max size display with STATE_ON, however, when folding the device and unfold quickly in a meeting state, there is a timing that all the displays are in STATE_OFF which causes the exception in bindToLifeCycle.

The fix is to return the max size display with STATE_OFF as a fallback.

Bug: b/255170076
Test: run the meet app on foldable device
Change-Id: Iab97302014182bdf722e879b8a66680ea376b75c